Episode 1 - Bloody Velen
The First Episode of the "True to the story" version of the Witcher TV Show Script Uploaded in both Docx and PDF for easy reading.

Make sure you check out the Preface and Readme, potentially to understand more, maybe to learn something new. It's worth it. These are all subject to change with development, these are literally 1st drafts written within a 72 hours period.